Refactor: Moved largest descendant computation
